Move download/upload functions to separate file
